requirements.txt
ase==3.22.1
basis_set_exchange==0.9.1
e3nn==0.4.4
hydra-core==1.3.2
ipdb==0.13.13
lightning==2.1.3
lmdb==1.4.1
lz4==4.3.2
numpy==1.26.4
omegaconf==2.3.0
p_tqdm==1.4.0
pandas==2.2.1
plotly==5.18.0
pymatgen==2023.11.12
python-dotenv==1.0.1
scipy==1.12.0
torch==1.13.1+cu116
torch_ema==0.3
tqdm==4.66.1
--extra-index-url https://download.pytorch.org/whl/cu116